s3:build: fix the build when no external talloc and tdb are installed.
The check for the external libs and the addition of the include paths
to the CPPFLAGS was too late in configure.

This patch moves the whole subsystem/library section up right below
the detection of "BLDSHARED".
And it updates not only SAMBA_CPPFLAGS but also SAMBA_CONFIGURE_CPPFLAGS
so that many tests that use these flags can now succeed.

Michael
(cherry picked from commit 06618b55e3e9d17cb7802aeed4769166118af373)