NSS wrapper - add dependency to "nsstest"
The code part makes use of the "NSS wrapper" if it is enabled. Surprisingly
the build process doesn't break with the missing dependency on more recent
systems. But with an older CentOS 4 installation it has been broken.
This patch should fix the problem on older distributions while keep the build
running on newer ones.